requirements.txt
openai==1.3.8
streamlit==1.39.0
pillow==10.4.0
python-dotenv==1.0.1
requests==2.32.3
httpx==0.24.1
